+++ App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import MainPage from './src/pages/MainPage';
-import { DateProvider } from './src/features/date/DateProvider';
+import { DateProvider } from './src/features/date/provider/DateProvider';
 
 export default function App(): JSX.Element {
   return (
