+++ src/pages/MainPage.tsx
@@ -18,8 +18,8 @@ import MainUser from '../assets/icon/mainuser.svg';
 import MainSearch from '../assets/icon/mainsearch.svg';
 import useFetch from '../shared/hooks/useFetch';
 import { getReprocessedIssues } from '../features/remakeissue/remotes/reprocessedissue';
-import { useDate } from '../features/date/DateProvider';
-import DateModal from '../dateSelect/components/DateModal';
+import { useDate } from '../features/date/provider/DateProvider';
+import DateModal from '../features/date/components/DateModal';
 import FollowUpIssueContainer from '../features/followupissue/components/FollowUpIssueContainer';
 import CategorySlider from '../features/remakeissue/components/CategorySlider';
 import { getFormatDate } from '../features/date/functions/formatDate';
